Split prepare phase into create and load phases. Keep prepare phase as (#13)
Split prepare phase into create and load phases. Keep prepare phase as well to be in line with sysbench's parent repo.

Create phase - creates the table and indexes
Load phase - loads the data

with this change, create phase has to be run with a single thread but the load phase can be run in parallel with multiple threads.